New folder/requirements.txt
discord.py
SpeechRecognition
pydub 